Path planning Robot
The terminology "optimal path planning" defines to the process of determining the safest, quickest, and the smoothest path between two points. The two points are the initial and final points. It is really important part in many aspect in robotics application. There are many application requires path planning with accuracy, safety and reliability. Many robotic applications, including as autonomous vehicles, surveillance operations, agricultural robots, and planetary and space exploration missions, require this duty. RRT* (Rapidly Exploring Random Tree Star) is a well-known sampling-based planning method. It has become really popular because of its dimensional complex problems. There are two methods used in this project to solve this problem; RRT and A star (*). Considering rapid demand of this field, this paper represents the development of an agent based on path planning approaches. Both the algorithms are compared and we’ll show the performance of both the algorithms